feat():direct lighting

+++ src/Common/math.hpp
@@ -20,12 +20,14 @@ typedef  glm::dmat3  mat3;
 typedef  glm::dmat4  mat4;
 #endif
 
+
+
 namespace Constant{
-    inline constexpr double PI = 3.14159265358979323846;
-    inline constexpr double INV_PI = 0.31830988618379067154;
-    inline constexpr double HALF_PI = 1.57079632679489661923;
-    inline constexpr double TWO_PI = 6.283185307179586476925;
-    inline constexpr double EPSILON = 1e-9;
+    inline constexpr Float PI = 3.14159265358979323846;
+    inline constexpr Float INV_PI = 0.31830988618379067154;
+    inline constexpr Float HALF_PI = 1.57079632679489661923;
+    inline constexpr Float TWO_PI = 6.283185307179586476925;
+    inline constexpr Float EPSILON = 1e-9;
 }
 
 inline  Float dot(vec3 a,vec3 b){
